requirements.txt
agno
python-dotenv
streamlit
openai
pymupdf
ddgs
cohere 
lancedb
tantivy
langchain
langchain-openai 
langchain-community

redis
redisvl